+++ R/revision_analysis.R
@@ -1,11 +1,13 @@
 #' A function to describe revision behavior for an archive.
 #'
 #' @description
-#' `revision_summary` removes all missing values (if requested), and then
-#'   computes some basic statistics about the revision behavior of an archive,
-#'   returning a tibble summarizing the revisions per time_value+epi_key
-#'   features. If `print_inform` is true, it prints a concise summary. The
-#'   columns returned are:
+#' `revision_summary` removes all missing values (if requested), and
+#'   then computes some basic statistics about the revision behavior
+#'   of an archive, returning a tibble summarizing the revisions per
+#'   time_value+epi_key features (limited to those that have data
+#'   available past the min waiting period to compare against, and are
+#'   not detected as added in a bulk report). The columns returned
+#'   are:
 #'  1. `n_revisions`: the total number of revisions for that entry
 #'  2. `min_lag`: the minimum time to any value (if `drop_nas=FALSE`, this
 #'   includes `NA`'s)
@@ -26,34 +28,60 @@
 #'   the window afterwards at 150.
 #'
 #' @param epi_arch an epi_archive to be analyzed
-#' @param ... <[`tidyselect`][dplyr::dplyr_tidy_select]>, used to choose the column to
-#'   summarize. If empty and there is only one value/measurement column (i.e.,
-#'   not in [`key_colnames`]) in the archive, it will automatically select it.
-#'   If supplied, `...` must select exactly one column.
-#' @param drop_nas bool, drop any `NA` values from the archive? After dropping
-#'   `NA`'s compactify is run again if `compactify` is `TRUE` to make
-#'   sure there are no duplicate values from occasions when the signal is
-#'   revised to `NA`, and then back to its immediately-preceding value.
-#' @param min_waiting_period `difftime`, integer or `NULL`. Sets a cutoff: any
-#'   time_values that have not had at least `min_waiting_period` to stabilize as
-#'   of the `versions_end` are removed. `min_waiting_period` should characterize
-#'   the typical time during which most significant revisions occur. The default
-#'   of 60 days corresponds to a typical near-final value for case counts as
-#'   reported in the context of insurance. To avoid this filtering, either set
-#'   to `NULL` or 0. A `difftime` will be rounded up to the appropriate `time_type` if
-#'   necessary (that is 5 days will be rounded to 1 week if the data is weekly).
-#' @param within_latest double between 0 and 1. Determines the threshold
-#'   used for the `lag_to`
-#' @param compactify bool. If `TRUE`, we will compactify after the signal
-#'   requested in `...` has been selected on its own and the `drop_nas` step.
-#'   This helps, for example, to give similar results when called on
-#'   [merged][epix_merge] and single-signal archives, since merged archives
-#'   record an update when any of the other signals change, not just the
-#'   requested signal. The default is `TRUE`.
-#' @param compactify_abs_tol length-1 double, used if `compactify` is `TRUE`, it
-#'   determines the threshold for when two doubles are considered identical.
-#' @param return_only_tibble boolean to return only the simple `tibble` of
-#'   computational results rather than the complete S3 object.
+#' @param ... <[`tidyselect`][dplyr::dplyr_tidy_select]>, used to
+#'   choose the column to summarize. If empty and there is only one
+#'   value/measurement column (i.e., not in [`key_colnames`]) in the
+#'   archive, it will automatically select it.  If supplied, `...`
+#'   must select exactly one column.
+#' @param min_waiting_period `difftime`, integer or `NULL`. Sets a
+#'   cutoff: any time_values that have not had at least
+#'   `min_waiting_period` to stabilize as of the `versions_end` are
+#'   removed. `min_waiting_period` should characterize the typical
+#'   time during which most significant revisions occur. The default
+#'   of 60 days corresponds to a typical near-final value for case
+#'   counts as reported in the context of insurance. To avoid this
+#'   filtering, either set to `NULL` or 0. A `difftime` will be
+#'   rounded up to the appropriate `time_type` if necessary (that is 5
+#'   days will be rounded to 1 week if the data is weekly).
+#' @param within_latest double between 0 and 1. Determines the
+#'   threshold used for the `lag_to`
+#' @param bulk_reporting_level,bulk_reporting_multiplier numeric; the
+#'   former between 0 and 1, typically close to but less than one, and
+#'   the latter `>= 1`; defaults of 0.8 and 1.2,
+#'   respectively. Determines how to detect bulk reporting.  Consider
+#'   the distribution of "max initial lags" across geodemographic
+#'   group x version pairs that add initial observations for new time
+#'   values; a bulk reporting lag threshold is determined by taking
+#'   the `bulk_reporting_level`-th quantile of this distribution,
+#'   multiplying by `bulk_reporting_multiplier`, and rounding to an
+#'   integer number of time intervals.  To avoid flagging anything as
+#'   bulk reporting, set `bulk_reporting_level = 1`.
+#' @param compactify bool. If `TRUE`, we will compactify after the
+#'   signal requested in `...` has been selected on its own and the
+#'   `drop_nas` step.  This helps, for example, to give similar
+#'   results when called on [merged][epix_merge] and single-signal
+#'   archives, since merged archives record an update when any of the
+#'   other signals change, not just the requested signal. The default
+#'   is `TRUE`.
+#' @param compactify_abs_tol length-1 double, used if `compactify` is
+#'   `TRUE`, it determines the threshold for when two doubles are
+#'   considered identical.
+#' @param compactify_drop_initial_nas bool; should we drop initial
+#'   estimates of NA during the compactification step?  Default is
+#'   TRUE, because these NAs are likely present due to
+#'   [`epix_merge()`]ing with a more timely indicator, and the
+#'   upstream source probably didn't actually report explicit NAs as
+#'   provisional estimates.
+#' @param drop_nas bool, do we drop all `NA` values (not just initial
+#'   estimates of `NA`) from the archive data structure prior to
+#'   (optional) compactification?  This both strips (i) initial
+#'   estimates of NA and (ii) explicit revisions from a non-NA
+#'   estimate to NA.  Equivalent to `compactify_drop_initial_nas` if
+#'   `compactify` is `TRUE`. Default is FALSE, favoring `compactify =
+#'   TRUE, compactify_drop_initial_nas = TRUE`.
+#' @param return_only_tibble boolean to return only the simple
+#'   `tibble` of computational results rather than the complete S3
+#'   object.
 #'
 #' @details Applies to `epi_archive`s with `time_type`s of `"day"`, `"week"`,
 #'   and `"yearmonth"`. It can also work with a `time_type` of `"integer"` if
@@ -63,13 +91,18 @@
 #'   produce incorrect results for some calculations, since week numbering
 #'   contains jumps at year boundaries.
 #'
-#' @return An S3 object with class `revision_behavior`. This function is typically
+#' @return An S3 object with class `revision_analysis`. This function is typically
 #'   called for the purposes of inspecting the printed output. The
 #'   results of the computations are available in
 #'   `revision_analysis(...)$revision_behavior`. If you only want to access
 #'   the internal computations, use `return_only_tibble = TRUE`.
 #'
 #' @examples
+#'
+#' # Print revision summary:
+#' revision_analysis(archive_cases_dv_subset, percent_cli)
+#'
+#' # Print some underlying data:
 #' revision_example <- revision_analysis(archive_cases_dv_subset, percent_cli)
 #' revision_example$revision_behavior %>% arrange(desc(spread))
 #'
@@ -77,15 +110,19 @@
 #' @importFrom cli cli_inform cli_abort cli_li
 #' @importFrom rlang list2 syms dots_n
 #' @importFrom vctrs vec_cast
+#' @importFrom stats quantile
 #' @importFrom dplyr mutate group_by arrange filter if_any all_of across pull pick c_across
 #'   everything ungroup summarize if_else %>%
 revision_analysis <- function(epi_arch,
                               ...,
-                              drop_nas = TRUE,
                               min_waiting_period = as.difftime(60, units = "days"),
                               within_latest = 0.2,
+                              bulk_reporting_level = 0.8,
+                              bulk_reporting_multiplier = 1.2,
                               compactify = TRUE,
                               compactify_abs_tol = 0,
+                              compactify_drop_initial_nas = TRUE,
+                              drop_nas = FALSE,
                               return_only_tibble = FALSE) {
   assert_class(epi_arch, "epi_archive")
   if (methods::is(min_waiting_period, "difftime")) {
@@ -130,6 +167,8 @@ revision_analysis <- function(epi_arch,
   time_type <- epi_arch$time_type
 
   revision_behavior <- epi_arch$DT %>%
+    as.data.frame() %>%
+    as_tibble() %>%
     select(all_of(unique(c(ukey_names, arg))))
   if (!is.null(min_waiting_period)) {
     last_semistable_time_value <- time_minus_n_steps(
@@ -140,58 +179,134 @@ revision_analysis <- function(epi_arch,
     revision_behavior <- revision_behavior %>%
       filter(time_value <= last_semistable_time_value)
   }
-
   if (drop_nas) {
     # if we're dropping NA's, we should recompactify
     revision_behavior <-
       revision_behavior %>%
       filter(!is.na(.data[[arg]]))
-  } else {
-    revision_behavior <- epi_arch$DT
   }
   if (compactify) {
     revision_behavior <- revision_behavior %>%
-      apply_compactify(ukey_names, compactify_abs_tol)
+      apply_compactify(ukey_names, compactify_abs_tol, init_nas_are_locf = compactify_drop_initial_nas)
   }
+  if (time_type == "week") {
+    lubridate_starting_iso_wday <- getOption("lubridate.week.start", 7)
+    starting_lt_wday <- lubridate_starting_iso_wday %% 7
+    ending_lt_wday <- (starting_lt_wday + 6) %% 7
+    revision_behavior <- revision_behavior %>%
+      mutate(lag = pmax(0, time_minus_time_in_n_steps(
+        version,
+        time_value + (ending_lt_wday - as.POSIXlt(time_value)$wday) %% 7,
+        time_type,
+        require_integer = FALSE
+      )))
+  } else {
+    revision_behavior <- revision_behavior %>%
+      mutate(lag = pmax(0, time_minus_time_in_n_steps(version, time_value, time_type, require_integer = FALSE)))
+  }
+
+  total_na <- revision_behavior %>%
+    filter(is.na(c_across(!!arg))) %>% # nolint: object_usage_linter
+    nrow()
+  n_obs <- nrow(revision_behavior)
+
+  # Note: `setDT()` modifies `revision_behavior` in-place, changing
+  # the object it points to from a tibble to a data.table for the
+  # remainder of this function, though we eventually point it to
+  # something that's a tibble later on.
+  setDT(revision_behavior)
+
+  initial_reporting <- revision_behavior %>%
+    #   We extract the initial report
+    # for each epikey-time combination using `.SD[1]` (relying on
+    # chronological sorting by version), then calculate the min/max lag
+    # for all first-time data introduced in each version dump.
+    .[, .SD[1], by = c(epikeytime_names)] %>%
+    .[, list(
+      # We suppress warnings because calling `min` or `max` on empty inputs (e.g.,
+      # during data.table prototype evaluation on a 0-row table) warns about no
+      # non-missing arguments and returns Inf / -Inf.
+      min_initial_lag = suppressWarnings(min(lag)),
+      max_initial_lag = suppressWarnings(max(lag))
+    ),
+    by = c(epikey_names, "version")
+    ] %>%
+    setDF() %>%
+    as_tibble()
+  assert_numeric(bulk_reporting_level, lower = 0, upper = 1, any.missing = FALSE, len = 1L)
+  assert_numeric(bulk_reporting_multiplier, lower = 1, any.missing = FALSE, len = 1L)
+  bulk_reporting_quantile <- unname(quantile(initial_reporting$max_initial_lag, bulk_reporting_level))
+  max_nonbulk_initial_lag <- round(bulk_reporting_multiplier * bulk_reporting_quantile)
+  bulk_reporting_versions <- initial_reporting %>%
+    filter(.data$max_initial_lag > .env$max_nonbulk_initial_lag) %>%
+    pull(version) %>%
+    vec_unique() %>%
+    vec_sort()
+  nonbulk_expanding_versions <-
+    vec_set_difference(initial_reporting$version, bulk_reporting_versions) %>%
+    vec_sort()
+  revision_only_versions <-
+    vec_set_difference(revision_behavior$version, c(bulk_reporting_versions, nonbulk_expanding_versions)) %>%
+    vec_sort()
+
+  data.table::setnames(revision_behavior, arg, ".VAL") # faster than .SD[[arg]] later on
+
   revision_behavior <-
     revision_behavior %>%
-    mutate(lag = time_minus_time_in_n_steps(version, time_value, time_type)) %>% # nolint: object_usage_linter
-    group_by(pick(all_of(epikeytime_names))) %>% # group = versions of one measurement
-    summarize(
-      n_revisions = dplyr::n() - 1,
-      min_lag = min(lag), # nolint: object_usage_linter
-      max_lag = max(lag), # nolint: object_usage_linter
-      min_value = f_no_na(min, .data[[arg]]),
-      max_value = f_no_na(max, .data[[arg]]),
-      median_value = f_no_na(median, .data[[arg]]),
-      lag_to = lag_within_x_latest(lag, .data[[arg]], prop = within_latest),
-      .groups = "drop"
-    ) %>%
+    {
+      with_missing <- .[, list(
+        n_revisions = .N - 1L,
+        # We suppress warnings because calling `min` or `max` on empty inputs
+        # warns about no non-missing arguments.
+        min_lag = suppressWarnings(min(lag)), # nolint: object_usage_linter
+        max_lag = suppressWarnings(max(lag)), # nolint: object_usage_linter
+        lag_to = lag_within_x_latest(lag, .VAL, prop = ..within_latest) # nolint: object_usage_linter
+      ), by = c(epikeytime_names)]
+      without_missing <- .[!is.na(.VAL)][, list(
+        min_value = suppressWarnings(min(.VAL)), # nolint: object_usage_linter
+        max_value = suppressWarnings(max(.VAL)), # nolint: object_usage_linter
+        median_value = suppressWarnings(median(.VAL)) # nolint: object_usage_linter
+      ), by = c(epikeytime_names)]
+      merge(with_missing, without_missing, by = epikeytime_names, all = TRUE)
+    } %>%
+    setDF() %>%
+    as_tibble()
+  n_epikeytimes <- nrow(revision_behavior)
+  revision_behavior <- revision_behavior %>% # nolint: object_usage_linter
+    filter(.data$min_lag <= .env$max_nonbulk_initial_lag) %>%
     mutate(
       spread = max_value - min_value, # nolint: object_usage_linter
       rel_spread = spread / max_value, # nolint: object_usage_linter
-      min_lag = n_steps_to_time_delta(min_lag, time_type), # nolint: object_usage_linter
-      max_lag = n_steps_to_time_delta(max_lag, time_type), # nolint: object_usage_linter
-      lag_near_latest = n_steps_to_time_delta(lag_to, time_type) # nolint: object_usage_linter
+      min_lag = n_steps_to_time_delta(min_lag, time_type, require_integer = FALSE), # nolint: object_usage_linter
+      max_lag = n_steps_to_time_delta(max_lag, time_type, require_integer = FALSE), # nolint: object_usage_linter
+      lag_near_latest = n_steps_to_time_delta(lag_to, time_type, require_integer = FALSE) # nolint: object_usage_linter
     ) %>%
     select(-lag_to) %>%
     relocate(
       time_value, geo_value, all_of(epikey_names), n_revisions, min_lag, max_lag, # nolint: object_usage_linter
       lag_near_latest, spread, rel_spread, min_value, max_value, median_value # nolint: object_usage_linter
     )
-  total_na <- epi_arch$DT %>%
-    filter(is.na(c_across(!!arg))) %>% # nolint: object_usage_linter
-    nrow()
+  n_nonbulk_epikeytimes <- nrow(revision_behavior)
   if (!return_only_tibble) {
-    revision_behavior <- structure(list(
+    revision_behavior <- structure(list( # nolint: object_usage_linter
       revision_behavior = revision_behavior,
+      initial_reporting = initial_reporting,
+      max_nonbulk_initial_lag = n_steps_to_time_delta( # nolint: object_usage_linter
+        max_nonbulk_initial_lag, time_type,
+        require_integer = FALSE
+      ),
+      bulk_reporting_versions = bulk_reporting_versions,
+      nonbulk_expanding_versions = nonbulk_expanding_versions,
+      revision_only_versions = revision_only_versions,
       range_time_values = range(epi_arch$DT$time_value),
       signal_variable = arg,
       drop_nas = drop_nas,
       time_type = time_type,
       total_na = total_na,
       max_val = max(epi_arch$DT[[arg]], na.rm = TRUE),
-      n_obs = nrow(epi_arch$DT),
+      n_obs = n_obs,
+      n_epikeytimes = n_epikeytimes,
+      n_nonbulk_epikeytimes = n_nonbulk_epikeytimes,
       within_latest = within_latest
     ), class = "revision_analysis")
   }
@@ -231,22 +346,46 @@ print.revision_analysis <- function(x,
   }
   if (is.null(abs_spread_threshold)) abs_spread_threshold <- .05 * x$max_val
   rev_beh <- x$revision_behavior
-  cli::cli_h2("An epi_archive spanning {.val {x$range_time_values[1]}} to {.val {x$range_time_values[2]}}.")
-  cli::cli_h3("Min lag (time to first version):")
-  time_delta_summary(rev_beh$min_lag, x$time_type) %>% print()
+  cli::cli_h2("Revision analysis for archive spanning time values {.val {x$range_time_values[1]}} to {.val {x$range_time_values[2]}}.") # nolint: line_length_linter
+  cli::cli_h3("Across epi_key + versions that add new time values:")
+  cli::cli_inform("Freshest new time value's lag/latency:")
+  time_delta_summary(x$initial_reporting$min_initial_lag, x$time_type) %>% print()
+  cli::cli_inform("Farthest-back new time value's lag/latency:")
+  time_delta_summary(x$initial_reporting$max_initial_lag, x$time_type) %>% print()
   if (!x$drop_nas) {
+    cli::cli_h3("Across epi_key + time_value + versions:")
     cli_inform("Fraction of all versions that are `NA`:")
     cli_li(num_percent(x$total_na, x$n_obs, ""))
-    cli_inform("")
   }
+
+  if (vec_size(x$bulk_reporting_versions) > 0L) {
+    cli::cli_h3("Bulk reporting adding initial observations for older epikey + time values:")
+    cli_inform("Initial lags above {format_time_delta(x$max_nonbulk_initial_lag, x$time_type)}
+                were counted as bulk reporting.")
+    cli_inform("Fraction of epi_key + time_values initially added by bulk reporting:")
+    cli_li(num_percent(x$n_epikeytimes - x$n_nonbulk_epikeytimes, x$n_epikeytimes, ""))
+    cli_inform("Versions containing bulk reporting: {vec_size(x$bulk_reporting_versions)}")
+    cli_li("({x$bulk_reporting_versions})")
+    cli_inform("Versions adding epikey + time values but no bulk reporting: {vec_size(x$nonbulk_expanding_versions)}")
+    cli_inform("Revision-only versions: {vec_size(x$revision_only_versions)}")
+  } else {
+    cli::cli_h3("Bulk reporting of older epikey + time values: none detected")
+    cli_inform("Initial lags above {format_time_delta(x$max_nonbulk_initial_lag, x$time_type)}
+                would have been counted as bulk reporting.")
+  }
+
+  # \u00a0 is non-breaking space to align text across lines here:
+  cli::cli_h3("Remaining information is for non-bulk-reported epikey + time values
+               \u00a0\u00a0\u00a0with semi-stable versions past the waiting period available.")
+
   cli::cli_h3("Fraction of epi_key + time_values with")
   total_num <- nrow(rev_beh) # nolint: object_usage_linter
   total_num_unrevised <- sum(rev_beh$n_revisions == 0) # nolint: object_usage_linter
   cli_inform("No revisions:")
   cli_li(num_percent(total_num_unrevised, total_num, ""))
   total_quickly_revised <- sum( # nolint: object_usage_linter
-    time_delta_to_n_steps(rev_beh$max_lag, x$time_type) <=
-      time_delta_to_n_steps(quick_revision, x$time_type)
+    time_delta_to_n_steps(rev_beh$max_lag, x$time_type, require_integer = FALSE) <=
+      time_delta_to_n_steps(quick_revision, x$time_type, require_integer = FALSE)
   )
   cli_inform("Quick revisions (last revision within {format_time_delta(quick_revision, x$time_type)}
                 of the `time_value`):")
@@ -275,9 +414,23 @@ print.revision_analysis <- function(x,
 
   # time_type_unit_pluralizer[[time_type]] is a format string controlled by us
   # and/or downstream devs, so we can paste it onto our format string safely:
-  units_plural <- pluralize(paste0("{qty(2)}", time_type_unit_pluralizer[[x$time_type]])) # nolint: object_usage_linter
-  cli::cli_h3("{toTitleCase(units_plural)} until within {.val {x$within_latest*100}}% of the latest value:")
+  lag_near_latest_time_type <- x$time_type
+  if (inherits(rev_beh[["lag_near_latest"]], "difftime")) {
+    lag_near_latest_time_type <- gsub("s$", "", units(rev_beh[["lag_near_latest"]]))
+  }
+  lag_near_latest_units_plural <- pluralize(paste0("{qty(2)}", time_type_unit_pluralizer[[lag_near_latest_time_type]])) # nolint: object_usage_linter
+  cli::cli_h3(
+    "{toTitleCase(lag_near_latest_units_plural)} until within {.val {x$within_latest*100}}% of the latest value:"
+  )
   time_delta_summary(rev_beh[["lag_near_latest"]], x$time_type) %>% print()
+
+  max_lag_time_type <- x$time_type
+  if (inherits(rev_beh[["max_lag"]], "difftime")) {
+    max_lag_time_type <- gsub("s$", "", units(rev_beh[["max_lag"]]))
+  }
+  max_lag_units_plural <- pluralize(paste0("{qty(2)}", time_type_unit_pluralizer[[max_lag_time_type]])) # nolint: object_usage_linter
+  cli::cli_h3("{toTitleCase(max_lag_units_plural)} until at the latest lag:")
+  time_delta_summary(rev_beh[["max_lag"]], x$time_type) %>% print()
 }
 
 #' @export
@@ -290,6 +443,10 @@ revision_summary <- revision_analysis
 #' @param prop optional length-1 double; proportion
 #' @keywords internal
 lag_within_x_latest <- function(lags, values, prop = .2) {
+  # Guard against empty inputs to prevent out-of-bounds indexing errors like `values[[0]]`.
+  if (length(values) == 0L) {
+    return(NA_real_)
+  }
   latest_value <- values[[length(values)]]
   close_enough <- abs(values - latest_value) < prop * latest_value
   # we want to ignore any stretches where it's close, but goes farther away later
@@ -335,18 +492,36 @@ num_percent <- function(a, b, b_description) {
 #' steps).
 #'
 #' @keywords internal
-time_delta_summary <- function(time_delta, time_type) {
+time_delta_summary <- function(time_delta, time_type = NULL) {
+  if (inherits(time_delta, "difftime")) {
+    time_type <- gsub("s$", "", units(time_delta))
+  }
+  if (is.null(time_type)) {
+    cli_abort("time_type must be provided if time_delta is not a difftime")
+  }
   if (length(time_delta) > 0) {
-    n_steps <- time_delta_to_n_steps(time_delta, time_type)
-    res <- data.frame(
+    n_steps <- time_delta_to_n_steps(time_delta, time_type, require_integer = FALSE)
+    summary_vals <- c(
       min = min(n_steps),
       median = median(n_steps),
-      mean = round(mean(n_steps), 1),
-      max = max(n_steps),
+      mean = mean(n_steps),
+      max = max(n_steps)
+    )
+    summary_deltas <- n_steps_to_time_delta(summary_vals, time_type, require_integer = FALSE)
+    res <- data.frame(
+      min = summary_deltas[1],
+      median = summary_deltas[2],
+      mean = summary_deltas[3],
+      max = summary_deltas[4],
       row.names = " ",
       check.names = FALSE
-    ) %>%
-      mutate(across(c(min, median, mean, max), ~ .x * unit_time_delta(time_type)))
+    )
+    if (inherits(res$mean, "difftime")) {
+      units_saved <- units(res$mean)
+      res$mean <- as.difftime(round(as.numeric(res$mean), 1), units = units_saved)
+    } else {
+      res$mean <- round(res$mean, 1)
+    }
     res
   } else {
     data.frame()